Vertexing: hide VertexingError implementation
