+++ Core/src/Vertexing/CMakeLists.txt
@@ -2,4 +2,5 @@ target_sources_local(
   ActsCore
   PRIVATE
     FsmwMode1dFinder.cpp
+    VertexingError.cpp
 )
