+++ Core/src/Vertexing/VertexingError.cpp
@@ -1,0 +1,46 @@
+// This file is part of the Acts project.
+//
+// Copyright (C) 2020 CERN for the benefit of the Acts project
+//
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+#include "Acts/Vertexing/VertexingError.hpp"
+
+namespace {
+
+class VertexingErrorCategory : public std::error_category {
+ public:
+  // Return a short descriptive name for the category.
+  const char* name() const noexcept final { return "KalmanFitterError"; }
+
+  // Return what each enum means in text.
+  std::string message(int c) const final {
+    using Acts::VertexingError;
+
+    switch (static_cast<VertexingError>(c)) {
+      case VertexingError::NumericFailure:
+        return "Numeric failure in calculation.";
+      case VertexingError::EmptyInput:
+        return "Empty input provided.";
+      case VertexingError::SeedingError:
+        return "Error while finding vertex seed.";
+      case VertexingError::NotConverged:
+        return "Unable to converge.";
+      case VertexingError::ElementNotFound:
+        return "Unable to find element.";
+      case VertexingError::NoCovariance:
+        return "No covariance provided.";
+      default:
+        return "unknown";
+    }
+  }
+};
+
+}  // namespace
+
+std::error_code Acts::make_error_code(Acts::VertexingError e) {
+  static VertexingErrorCategory c;
+  return {static_cast<int>(e), c};
+}
